Fixed variant support updating to only try to set when the table does exist.

+++ src/system/modules/metamodels/MetaModelTableManipulation.php
@@ -20,7 +20,7 @@ if (!defined('TL_ROOT'))
 
 /**
  * This is the class for table manipulations like creation/renaming/deleting of tables and columns.
- * 
+ *
  * @package	   MetaModel
  * @subpackage Core
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
@@ -84,7 +84,7 @@ class MetaModelTableManipulation
 
 	/**
 	 * Returns the Contao database instance to use.
-	 * 
+	 *
 	 * @return Database the database instance.
 	 */
 	protected static function getDB()
@@ -94,9 +94,9 @@ class MetaModelTableManipulation
 
 	/**
 	 * Checks wheter the given table name is valid.
-	 * 
+	 *
 	 * @param string $strTableName the table name to check
-	 * 
+	 *
 	 * @return bool true if the table name is valid, false otherwise.
 	 */
 	public static function isValidTablename($strTableName)
@@ -108,9 +108,9 @@ class MetaModelTableManipulation
 
 	/**
 	 * Checks whether the column with the given name is a MetaModel system column.
-	 * 
+	 *
 	 * @param string $strColName the name of the column
-	 * 
+	 *
 	 * @return bool true if the column is a system column, false otherwise.
 	 */
 	public static function isValidColumnName($strColName)
@@ -120,9 +120,9 @@ class MetaModelTableManipulation
 
 	/**
 	 * Checks whether the column with the given name is a MetaModel system column.
-	 * 
+	 *
 	 * @param string $strColName the name of the column
-	 * 
+	 *
 	 * @return bool true if the column is a system column, false otherwise.
 	 */
 	public static function isSystemColumn($strColName)
@@ -133,9 +133,9 @@ class MetaModelTableManipulation
 	/**
 	 * Checks whether the given table name is valid.
 	 * Throws an Exception if the table name is invalid.
-	 * 
+	 *
 	 * @param string $strTableName the table name to check
-	 * 
+	 *
 	 * @return void
 	 */
 	public static function checkTablename($strTableName)
@@ -147,12 +147,12 @@ class MetaModelTableManipulation
 	}
 
 	/**
-	 * Checks whether the column with the given name is not a MetaModel system column and is a valid Database 
+	 * Checks whether the column with the given name is not a MetaModel system column and is a valid Database
 	 * column name, @see{MetaModelTableManipulation::isSystemColumn()} and @see{MetaModelTableManipulation::isValidColumnName()}.
 	 * If there is any problem, an Exception is raised, stating the nature of the error in the Exception message.
-	 * 
+	 *
 	 * @param string $strColName the name of the column
-	 * 
+	 *
 	 * @return void
 	 */
 	public static function checkColumnName($strColName)
@@ -171,9 +171,9 @@ class MetaModelTableManipulation
 	/**
 	 * Checks whether the given table exists.
 	 * Throws an Exception if the table name is invalid or does not exist.
-	 * 
+	 *
 	 * @param string $strTableName the table name to check
-	 * 
+	 *
 	 * @return void
 	 */
 	public static function checkTableExists($strTableName)
@@ -181,16 +181,16 @@ class MetaModelTableManipulation
 		self::checkTablename($strTableName);
 		if (!self::getDB()->tableExists($strTableName, null, true))
 		{
-			throw new Exception(sprintf($GLOBALS['TL_LANG']['ERR']['tableDoesNotExist'], $strTableName)); 
+			throw new Exception(sprintf($GLOBALS['TL_LANG']['ERR']['tableDoesNotExist'], $strTableName));
 		}
 	}
 
 	/**
 	 * ensures that the given table does not exist.
 	 * Throws an Exception if the table name is invalid or does exist.
-	 * 
+	 *
 	 * @param string $strTableName the table name to check
-	 * 
+	 *
 	 * @return void
 	 */
 	public static function checkTableDoesNotExist($strTableName)
@@ -198,16 +198,16 @@ class MetaModelTableManipulation
 		self::checkTablename($strTableName);
 		if (self::getDB()->tableExists($strTableName, null, true))
 		{
-			throw new Exception(sprintf($GLOBALS['TL_LANG']['ERR']['tableExists'], $strTableName)); 
+			throw new Exception(sprintf($GLOBALS['TL_LANG']['ERR']['tableExists'], $strTableName));
 		}
 	}
 
 	/**
 	 * Creates a table with the given name.
 	 * Throws Exception if the table name is invalid or already exists.
-	 * 
+	 *
 	 * @param string $strTableName the name of the new table to create.
-	 * 
+	 *
 	 * @return void
 	 */
 	public static function createTable($strTableName)
@@ -219,11 +219,11 @@ class MetaModelTableManipulation
 	/**
 	 * Renames a table with the given name to the given new name.
 	 * Throws Exception if the new table name is invalid.
-	 * 
+	 *
 	 * @param string $strTableName    the name of the table to rename.
-	 * 
+	 *
 	 * @param string $strNewTableName the name to which the table shall be renamed to.
-	 * 
+	 *
 	 * @return void
 	 * 	 */
 	public static function renameTable($strTableName, $strNewTableName)
@@ -237,9 +237,9 @@ class MetaModelTableManipulation
 	/**
 	 * Deletes the table with the given name.
 	 * Throws Exception if the table name is invalid or the table does not exist.
-	 * 
+	 *
 	 * @param string $strTableName the name of the new table to delete.
-	 * 
+	 *
 	 * @return void
 	 */
 	public static function deleteTable($strTableName)
@@ -252,12 +252,12 @@ class MetaModelTableManipulation
 	/**
 	 * Checks whether the given table exists.
 	 * Throws an Exception if the table name is invalid or does not exist.
-	 * 
+	 *
 	 * @param string $strTableName the table name to check
-	 * 
+	 *
 	 * @param string $strColName   the column name to check
-	 * 
-	 * 
+	 *
+	 *
 	 * @return void
 	 */
 	public static function checkColumnExists($strTableName, $strColName)
@@ -266,19 +266,19 @@ class MetaModelTableManipulation
 		self::checkColumnName($strTableName);
 		if (!self::getDB()->fieldExists($strColName, $strTableName, true))
 		{
-			throw new Exception(sprintf($GLOBALS['TL_LANG']['ERR']['columnDoesNotExist'], $strColName, $strTableName)); 
+			throw new Exception(sprintf($GLOBALS['TL_LANG']['ERR']['columnDoesNotExist'], $strColName, $strTableName));
 		}
 	}
 
 	/**
 	 * Checks whether the given table exists.
 	 * Throws an Exception if the table name is invalid or does not exist.
-	 * 
+	 *
 	 * @param string $strTableName the table name to check
-	 * 
+	 *
 	 * @param string $strColName   the column name to check
-	 * 
-	 * 
+	 *
+	 *
 	 * @return void
 	 */
 	public static function checkColumnDoesNotExist($strTableName, $strColName)
@@ -287,7 +287,7 @@ class MetaModelTableManipulation
 		self::checkColumnName($strTableName);
 		if (self::getDB()->fieldExists($strColName, $strTableName, true))
 		{
-			throw new Exception(sprintf($GLOBALS['TL_LANG']['ERR']['columnExists'], $strColName)); 
+			throw new Exception(sprintf($GLOBALS['TL_LANG']['ERR']['columnExists'], $strColName));
 		}
 	}
 
@@ -295,16 +295,16 @@ class MetaModelTableManipulation
 
 	/**
 	 * Add a column to a table.
-	 * 
+	 *
 	 * Throws Exception if the table does not exist, the column name is invalid or the column already exists.
-	 * 
+	 *
 	 * @param string $strTableName  the name of the table to add the column to.
-	 * 
+	 *
 	 * @param string $strColumnName the name of the new column.
-	 * 
+	 *
 	 * @param string $strType       the SQL type notation of the new column.
-	 * 
-	 * 
+	 *
+	 *
 	 * @return void
 	 */
 	public static function createColumn($strTableName, $strColumnName, $strType)
@@ -323,18 +323,18 @@ class MetaModelTableManipulation
 
 	/**
 	 * Rename a column and/or change it's type in a table.
-	 * 
+	 *
 	 * Throws Exception if the table does not exist, the column name is invalid or the column already exists.
-	 * 
+	 *
 	 * @param string $strTableName     the name of the table the column is in.
-	 * 
+	 *
 	 * @param string $strColumnName    the current name of the column to be renamed.
-	 * 
+	 *
 	 * @param string $strNewColumnName the new name for the column.
-	 * 
+	 *
 	 * @param string $strNewType       the new SQL type notation of the column.
-	 * 
-	 * 
+	 *
+	 *
 	 * @return void
 	 */
 	public static function renameColumn($strTableName, $strColumnName, $strNewColumnName, $strNewType)
@@ -359,11 +359,11 @@ class MetaModelTableManipulation
 	/**
 	 * Delete a column from a table.
 	 * Throws Exception if the table does not exist, the column name is invalid or the column does not exist.
-	 * 
+	 *
 	 * @param string $strTableName  the name of the table the column is in.
-	 * 
+	 *
 	 * @param string $strColumnName the name of the column to drop.
-	 * 
+	 *
 	 * @return void
 	 */
 	public static function dropColumn($strTableName, $strColumnName)
@@ -372,8 +372,8 @@ class MetaModelTableManipulation
 		// TODO: throw exceptions
 		self::getDB()->execute(
 						sprintf(
-							self::STATEMENT_DROP_COLUMN, 
-							$strTableName, 
+							self::STATEMENT_DROP_COLUMN,
+							$strTableName,
 							$strColumnName
 						)
 		);
@@ -381,18 +381,18 @@ class MetaModelTableManipulation
 
 	/**
 	 * Enables or disables Variant support on a certain MetaModel table.
-	 * 
+	 *
 	 * @param string $strTableName      the table name of the MetaModel.
-	 * 
+	 *
 	 * @param bool   $blnVariantSupport flag if the support shall be turned on or off.
-	 * 
+	 *
 	 * @return void
 	 */
 	public static function setVariantSupport($strTableName, $blnVariantSupport)
 	{
 		if ($blnVariantSupport)
 		{
-			if (!self::getDB()->fieldExists('varbase', $strTableName, true))
+			if (self::getDB()->tableExists($strTableName, null, true) && (!self::getDB()->fieldExists('varbase', $strTableName, true)))
 			{
 				self::createColumn($strTableName, 'varbase', 'char(1) NOT NULL default \'\'');
 				self::createColumn($strTableName, 'vargroup', 'int(11) NOT NULL default 0');
@@ -403,7 +403,7 @@ class MetaModelTableManipulation
 				self::getDB()->execute(sprintf('UPDATE %s SET vargroup=id, varbase=1', $strTableName));
 			}
 		} else {
-			if (self::getDB()->fieldExists('varbase', $strTableName, true))
+			if (self::getDB()->tableExists($strTableName, null, true) && self::getDB()->fieldExists('varbase', $strTableName, true))
 			{
 				self::dropColumn($strTableName, 'varbase');
 				self::dropColumn($strTableName, 'vargroup');
